Airflow/requirements.txt
apache-airflow-providers-apache-hdfs==4.11.1
apache-airflow-providers-apache-spark==5.4.1
apache-airflow-providers-common-compat==1.10.1

# Climate data ingestion dependencies
cdsapi==0.7.4
netCDF4==1.7.2
numpy==1.26.4